Fraction of mileage driven with a cold engine or catalizer below normal temperature

This function depends length of trip and on ambient temperature. From the guidelines EMEP/EEA air pollutant emission inventory guidebook http://www.eea.europa.eu/themes/air/emep-eea-air-pollutant-emission-inventory-guidebook

cold_mileage(ltrip, ta)

Arguments

ltrip
Numeric; Length of trip. It must be in 'units' km.

ta
Numeric or data.frame; average monthly temperature Celsius. It if is a data.frame, it is convenient that each column is each month.

Note

This function is set so that values vaires between 0 and 1.

Examples

if (FALSE) {
lkm <- units::set_units(1:10, km)
ta <- celsius(matrix(0:9, ncol = 12, nrow = 10))
a <- cold_mileage(lkm, rbind(ta, ta))
(a)
filled.contour(as.matrix(a), col = cptcity::lucky(n = 16))
}