Emissions factors for Environment Company of Sao Paulo, Brazil (CETESB) 2017

ef_cetesb returns a vector or data.frame of Brazilian emission factors.

ef_cetesb(p, veh, year = 2017, agemax = 40, full = FALSE,
  project = "constant")

Arguments

p
Character; Pollutants: "COd", "HCd", "NMHCd", "CH4", "NOxd", "CO2", "PM", "N2O", "KML", "FC", "NO2d", "NOd", "gD/KWH", "gCO2/KWH", "RCHOd", "CO", "HC", "NMHC", "NOx", "NO2" ,"NO", "RCHO" (g/km). The letter 'd' means deteriorated factor. Also, evaporative emissions at average temperature ranges: "D_20_35", "S_20_35", "R_20_35", "D_10_25", "S_10_25", "R_10_25", "D_0_15", "S_0_15" and "R_0_15" where D means diurnal (g/day), S hot/warm soak (g/trip) and R hot/warm running losses (g/trip).

veh
Character; Vehicle categories: "PC_G", "PC_FG", "PC_FE", "PC_E", "LCV_G", "LCV_FG", "LCV_FE", "LCV_E", "LCV_D", "SLT", "LT", "MT", "SHT", "HT", "UB", "SUB", "COACH", "ARTIC", "M_G_150", "M_G_150_500", "M_G_500", "M_FG_150", "M_FG_150_500", "M_FG_500", "M_FE_150", "M_FE_150_500", "M_FE_500", "CICLOMOTOR", "GNV"

year
Numeric; Filter the emission factor to start from a specific base year. If project is 'constant' values above 2017 and below 1980 will be repeated

agemax
Integer; age of oldest vehicles for that category

full
Logical; To return a data.frame instead or a vector adding Age, Year, Brazilian emissions standards and its euro equivalents.

project
haracter showing the method for projecting emission factors in future. Currently the only value is "constant"

Value

A vector of Emission Factor or a data.frame

Note

This emission factors are not exactly the same as the report of CETESB.

1) In this emission factors, there is also NO and NO2 based on split by published in the EMEP/EEA air pollutant emission inventory guidebook.

2) Also, the emission factors were extended till 50 years of use, repeating the oldest value.

3) CNG emission factors were expanded to other pollutants by comparison of US.EPA-AP42 emission factor: Section 1.4 Natural Gas Combustion.

References

Emissoes Veiculares no Estado de Sao Paulo 2016. Technical Report. url: https://cetesb.sp.gov.br/veicular/relatorios-e-publicacoes/.

Examples

{
a <- ef_cetesb("CO", "PC_G")
a <- ef_cetesb("R_10_25", "PC_G")
a <- ef_cetesb("CO", c("PC_G", "PC_FE"))
ef_cetesb(p = "CO", veh = "PC_G", year = 2018, agemax = 40)
ef_cetesb(p = "CO", veh = "PC_G", year = 1970, agemax = 40)
ef_cetesb(p = "CO", veh = "PC_G", year = 2030, agemax = 40)
}

#> Units: [g/trip]

#> Units: [g/km]
#>  [1]  0.141  0.141  0.141  0.141  0.141  0.141  0.141  0.141  0.141  0.141
#> [11]  0.141  0.141  0.141  0.141  0.114  0.155  0.211  0.241  0.274  0.274
#> [21]  0.204  0.199  0.369  0.302  0.302  0.340  0.350  0.400  0.430  0.480
#> [31]  0.730  0.740  0.790  1.200  3.800  4.700  6.000  6.300  6.200 11.500