Evaporative emission factor

ef_evap is a lookup table with tier 2 evaporative emission factors from EMEP/EEA emisison guidelines

ef_evap(ef, v, cc, dt, ca, pollutant = "NMHC", k = 1, ltrip, kmday,
  show = FALSE, verbose = FALSE)

Arguments

ef
Name of evaporative emission factor as *eshotc*: mean hot-soak with carburator, *eswarmc*: mean cold and warm-soak with carburator, eshotfi: mean hot-soak with fuel injection, *erhotc*: mean hot running losses with carburator, *erwarmc* mean cold and warm running losses, *erhotfi* mean hot running losses with fuel injection. Length of ef 1.

v
Type of vehicles, "PC", "Motorcycle", "Motorcycle_2S" and "Moped"

cc
Size of engine in cc. PC "<=1400", "1400_2000" and ">2000" Motorcycle_2S: "<=50". Motorcyces: ">50", "<=250", "250_750" and ">750". Only engines of >750 has canister.

dt
Character or Numeric: Average monthly temperature variation: "-5_10", "0_15", "10_25" and "20_35". This argument can vector with several elements. dt can also be data.frame, but it is recommended that the number of columns are each month. So that dt varies in each row and each column.

ca
Size of canister: "no" meaning no canister, "small", "medium" and "large".

pollutant
Character indicating any of the covered pollutants: "NMHC", "ethane", "propane", "i-butane", "n-butane", "i-pentane", "n-pentane", "2-methylpentane", "3-methylpentane", "n-hexane", "n-heptane", "propene", "trans-2-butene", "isobutene", "cis-2-butene", "1,3-butadiene", "trans-2-pentene", "cis-2-pentene", "isoprene", "propyne", "acetylene", "benzene", "toluene", "ethylbenzene", "m-xylene", "o-xylene", "1,2,4-trimethylbenzene" and "1,3,5-trimethylbenzene". Default is "NMHC"

k
multiplication factor

ltrip
Numeric; Length of trip. Experimental feature to conter g/trip and g/proced (assuming proced similar to trip) in g/km.

kmday
Numeric; average daily mileage. Experimental option to convert g/day in g/km. it is an information more solid than to know the average number of trips per day.

show
when TRUE shows row of table with respective emission factor.

verbose
Logical; To show more information

Value

emission factors in g/trip or g/proced. The object has class (g) but it order to know it is g/trip or g/proceed the argument show must by T

Note

Diurnal loses occur with daily temperature variations. Running loses occur during vehicles use. Hot soak emission occur following vehicles use.

References

Mellios G and Ntziachristos 2016. Gasoline evaporation. In: EEA, EMEP. EEA air pollutant emission inventory guidebook-2009. European Environment Agency, Copenhagen, 2009

Examples

{
# Do not run
ef_evap(ef = "eshotc", v = "PC", cc = "<=1400", dt = "0_15", ca = "no",
pollutant = "cis-2-pentene")
ef_evap(ef = "ed", v = "PC", cc = "<=1400", dt = "0_15", ca = "no",
show = TRUE)
ef_evap(ef = c("erhotc", "erhotc"), v = "PC", cc = "<=1400",
dt = "0_15", ca = "no",
show = TRUE)
ef_evap(ef = c("erhotc", "erhotc"), v = "PC", cc = "<=1400",
 dt = "0_15", ca = "no",
show = FALSE)
ef_evap(ef = "eshotc", v = "PC", cc = "<=1400", dt = "0_15", ca = "no",
show = TRUE)
ef_evap(ef = "erhotc", v = "PC", cc = "<=1400", dt = "0_15", ca = "no",
show = TRUE)
temps <- 10:20
ef_evap(ef = "erhotc", v = "PC", cc = "<=1400", dt = temps, ca = "no",
show = TRUE)
dt <- matrix(rep(1:24,5), ncol = 12) # 12 months
dt <- celsius(dt)
ef_evap(ef ="erhotc", v = "PC", cc = "<=1400",
dt = dt, ca = "no")
lkm <- units::set_units(10, km)
ef_evap(ef ="erhotc", v = "PC", cc = "<=1400", ltrip = lkm,
dt = dt, ca = "no")
}

#>               ef cols_temp
#> 1   0.167 [g/km]         1
#> 2   0.167 [g/km]         1
#> 3   0.239 [g/km]         1
#> 4   0.239 [g/km]         1
#> 5   0.239 [g/km]         1
#> 6   0.239 [g/km]         1
#> 7   0.239 [g/km]         1
#> 8   0.325 [g/km]         1
#> 9   0.325 [g/km]         1
#> 10  0.325 [g/km]         1
#> 11  0.325 [g/km]         2
#> 12  0.325 [g/km]         2
#> 13  0.325 [g/km]         2
#> 14  0.325 [g/km]         2
#> 15  0.325 [g/km]         2
#> 16  0.325 [g/km]         2
#> 17  0.325 [g/km]         2
#> 18  0.542 [g/km]         2
#> 19  0.542 [g/km]         2
#> 20  0.542 [g/km]         2
#> 21  0.542 [g/km]         3
#> 22  0.542 [g/km]         3
#> 23  0.542 [g/km]         3
#> 24  0.542 [g/km]         3
#> 25  0.167 [g/km]         3
#> 26  0.167 [g/km]         3
#> 27  0.239 [g/km]         3
#> 28  0.239 [g/km]         3
#> 29  0.239 [g/km]         3
#> 30  0.239 [g/km]         3
#> 31  0.239 [g/km]         4
#> 32  0.325 [g/km]         4
#> 33  0.325 [g/km]         4
#> 34  0.325 [g/km]         4
#> 35  0.325 [g/km]         4
#> 36  0.325 [g/km]         4
#> 37  0.325 [g/km]         4
#> 38  0.325 [g/km]         4
#> 39  0.325 [g/km]         4
#> 40  0.325 [g/km]         4
#> 41  0.325 [g/km]         5
#> 42  0.542 [g/km]         5
#> 43  0.542 [g/km]         5
#> 44  0.542 [g/km]         5
#> 45  0.542 [g/km]         5
#> 46  0.542 [g/km]         5
#> 47  0.542 [g/km]         5
#> 48  0.542 [g/km]         5
#> 49  0.167 [g/km]         5
#> 50  0.167 [g/km]         5
#> 51  0.239 [g/km]         6
#> 52  0.239 [g/km]         6
#> 53  0.239 [g/km]         6
#> 54  0.239 [g/km]         6
#> 55  0.239 [g/km]         6
#> 56  0.325 [g/km]         6
#> 57  0.325 [g/km]         6
#> 58  0.325 [g/km]         6
#> 59  0.325 [g/km]         6
#> 60  0.325 [g/km]         6
#> 61  0.325 [g/km]         7
#> 62  0.325 [g/km]         7
#> 63  0.325 [g/km]         7
#> 64  0.325 [g/km]         7
#> 65  0.325 [g/km]         7
#> 66  0.542 [g/km]         7
#> 67  0.542 [g/km]         7
#> 68  0.542 [g/km]         7
#> 69  0.542 [g/km]         7
#> 70  0.542 [g/km]         7
#> 71  0.542 [g/km]         8
#> 72  0.542 [g/km]         8
#> 73  0.167 [g/km]         8
#> 74  0.167 [g/km]         8
#> 75  0.239 [g/km]         8
#> 76  0.239 [g/km]         8
#> 77  0.239 [g/km]         8
#> 78  0.239 [g/km]         8
#> 79  0.239 [g/km]         8
#> 80  0.325 [g/km]         8
#> 81  0.325 [g/km]         9
#> 82  0.325 [g/km]         9
#> 83  0.325 [g/km]         9
#> 84  0.325 [g/km]         9
#> 85  0.325 [g/km]         9
#> 86  0.325 [g/km]         9
#> 87  0.325 [g/km]         9
#> 88  0.325 [g/km]         9
#> 89  0.325 [g/km]         9
#> 90  0.542 [g/km]         9
#> 91  0.542 [g/km]        10
#> 92  0.542 [g/km]        10
#> 93  0.542 [g/km]        10
#> 94  0.542 [g/km]        10
#> 95  0.542 [g/km]        10
#> 96  0.542 [g/km]        10
#> 97  0.167 [g/km]        10
#> 98  0.167 [g/km]        10
#> 99  0.239 [g/km]        10
#> 100 0.239 [g/km]        10
#> 101 0.239 [g/km]        11
#> 102 0.239 [g/km]        11
#> 103 0.239 [g/km]        11
#> 104 0.325 [g/km]        11
#> 105 0.325 [g/km]        11
#> 106 0.325 [g/km]        11
#> 107 0.325 [g/km]        11
#> 108 0.325 [g/km]        11
#> 109 0.325 [g/km]        11
#> 110 0.325 [g/km]        11
#> 111 0.325 [g/km]        12
#> 112 0.325 [g/km]        12
#> 113 0.325 [g/km]        12
#> 114 0.542 [g/km]        12
#> 115 0.542 [g/km]        12
#> 116 0.542 [g/km]        12
#> 117 0.542 [g/km]        12
#> 118 0.542 [g/km]        12
#> 119 0.542 [g/km]        12
#> 120 0.542 [g/km]        12