Base emissions factors from International Vehicle Emissions (IVE) model

ef_ive returns the base emission factors from the the IVE model. This function depend on vectorized mileage, which means your can enter with the mileage by age of use and the name of the pollutant.

ef_ive(description = "Auto/Sml Truck", fuel = "Petrol",
  weight = "Light", air_fuel_control = "Carburetor",
  exhaust = "None", evaporative = "PCV", mileage, pol,
  details = FALSE)

Arguments

description
Character; "Auto/Sml Truck" "Truck/Bus" or"Sml Engine".

fuel
Character; "Petrol", "NG Retrofit", "Natural Gas", "Prop Retro.", "Propane", "EthOH Retrofit", "OEM Ethanol", "Diesel", "Ethanol" or "CNG/LPG".

weight
Character; "Light", "Medium", "Heavy", "Lt", "Med" or "Hvy"

air_fuel_control
Character; One of the following characters: "Carburetor", "Single-Pt FI", "Multi-Pt FI", "Carb/Mixer", "FI", "Pre-Chamber Inject.", "Direct Injection", "2-Cycle", "2-Cycle, FI", "4-Cycle, Carb", "4-Cycle, FI" "4-Cycle"

exhaust
Character: "None", "2-Way", "2-Way/EGR", "3-Way", "3-Way/EGR", "None/EGR", "LEV", "ULEV", "SULEV", "EuroI", "EuroII", "EuroIII", "EuroIV", "Hybrid", "Improved", "EGR+Improv", "Particulate", "Particulate/NOx", "EuroV", "High Tech" or "Catalyst"

evaporative
Character: "PCV", "PCV/Tank" or"None".

mileage
Numeric; mileage of vehicle by age of use km.

pol
Character; One of the following characters: "Carburetor", "Single-Pt FI", "Multi-Pt FI", "Carb/Mixer", "FI", "Pre-Chamber Inject.", "Direct Injection", "2-Cycle", "2-Cycle, FI", "4-Cycle, Carb", "4-Cycle, FI" "4-Cycle" #'

"VOC_gkm"	"CO_gkm"	"NOx_gkm"

"PM_gkm"		"Pb_gkm"	"SO2_gkm"

"NH3_gkm"	"1,3-butadiene_gkm"		"formaldehyde_gkm"

"acetaldehyde_gkm"	"benzene_gkm"	"EVAP_gkm"

"CO2_gkm"	"N20_gkm"	"CH4_gkm"	"VOC_gstart"

"CO_gstart"	"NOx_gstart"	"PM_gstart"

"Pb_gstart"		"SO2_gstart"	"NH3_gstart"

"1,3-butadiene_gstart"	"formaldehyde_gstart"		"acetaldehyde_gstart"

"benzene_gstart"	"EVAP_gstart"	"CO2_gstart"

"N20_gstart"	"CH4_gstart"

"VOC_gkm"	"CO_gkm"

details
Logical; option to see or not more information about vehicle.

Value

An emission factor by annual mileage.

References

Nicole Davis, James Lents, Mauricio Osses, Nick Nikkila, Matthew Barth. 2005. Development and Application of an International Vehicle Emissions Model. Transportation Research Board, 81st Annual Meeting, January 2005, Washington, D.C.

Examples

{
# Do not run
# Passenger Cars PC
data(fkm)
# cumulative mileage from 1 to 50 years of use, 40:50
mil <- cumsum(fkm$KM_PC_E25(1:50))
ef_ive("Truck/Bus", mileage = mil, pol = "CO_gkm")
ef_ive(mileage = mil, pol = "CO_gkm", details = TRUE)
}

#> [1] "Smaller no control (US Pre-1972 style) light duty vehicle. Engine less than 1.5 liter and typically a sedan or smaller GVWR <5000 lbs."

#> Units: [g/km]
#>  [1] 22.255 22.255 22.255 22.255 22.255 27.140 27.140 27.140 27.140 27.140
#> [11] 27.140 34.221 34.221 34.221 34.221 34.221 34.221 34.221 34.221 34.221
#> [21] 34.221 34.221 34.221 34.221 34.221 34.221 34.221 34.221 34.221 34.221
#> [31] 34.221 34.221 34.221 34.221 34.221 34.221 34.221 34.221 34.221 34.221
#> [41] 34.221 34.221 34.221 34.221 34.221 34.221 34.221 34.221 34.221 34.221