Local Emissions factors

ef_local process an data.frame delivered by the user, but adding similar funcionality and arguments as ef_cetesb, which are classification, filtering and projections

ef_local(
  p,
  veh,
  year = 2017,
  agemax = 40,
  ef,
  full = FALSE,
  project = "constant",
  verbose = TRUE
)

Arguments

p
Character; pollutant delivered by the user. the name of the column of the data.frame must be Pollutant.

veh
Character; Vehicle categories available in the data.frame provided by the user

year
Numeric; Filter the emission factor to start from a specific base year. If project is 'constant' values above 2017 and below 1980 will be repeated

agemax
Integer; age of oldest vehicles for that category

ef
data.frame, for local the emission factors. The names of the ef must be `Age` `Year` `Pollutant` and all the vehicle categories...

full
Logical; To return a data.frame instead or a vector adding Age, Year, Brazilian emissions standards and its euro equivalents.

project
Character showing the method for projecting emission factors in future. Currently the only value is "constant"

verbose
Logical; To show more information

Value

A vector of Emission Factor or a data.frame

Details

returns a vector or data.frame of Brazilian emission factors.

Note

The names of the ef must be `Age` `Year` `Pollutant` and all the vehicle categories...

See also

Examples

if (FALSE) {
#do not run
}