Aggregate emissions by lumped groups in chemical mechanism

emis_chem aggregates emissions by chemical mechanism and convert grams to mol. This function reads all hydrocarbos and respective criteria polluants specified in ef_ldv_speed and ef_hdv_speed.

emis_chem(dfe, mechanism, colby, long = FALSE)

Arguments

dfe
data.frame with column `emissions` in grams and `pollutant` in long format.

mechanism
Character, any of "SAPRC", "RACM", "RADM2", "CBMZ", "MOZART", "SAPRC99", "CB05", "CB05CMAQ", "CB06CMAQ", "RACM2CMAQ", "SAPRC99CMAQ", "SAPRC07CMAQ", "SAPRC07A","CRIMECH" WRF: "RADM2_SORG", "CBMZ_MOSAIC", "CPTEC", "GOCART_CPTEC", "MOZEM", "MOZCEM", "CAMMAM", "MOZMEM", "MOZC_T1_EM", "CB05_OPT1" or "CB05_OPT2"

colby
Character indicating column name for aggregating extra column. For instance, region or province

long
Logical. Do you want data in long format?

Value

data.frame with lumped groups by chemical mechanism. It transform emissions in grams to mol.

Note

This feature is experimental and the mapping of pollutants and lumped species may change in future. This function is converting the intial data.frame input into data.table. To have a comprehensive speciation is necessary enter with a data.frame with colum 'emission' in long format including another column named 'pollutant' with species of NMHC, CO, NO, NO2, NH3, SO2, PM2.5 and coarse PM10.

Groups derived from gases has units 'mol' and from aersols 'g'. The aersol units for WRF-Chem are ug/m^2/s while for CMAQ and CAMx are g/s. So, leaving the units just in g, allow to make further change while providing flexibility for several models.

See also

Examples

{
# CO
df <- data.frame(emission = Emissions(1:10))
df$pollutant = "CO"
emis_chem(df, "CBMZ")
# hexanal
df$pollutant = "hexanal"
emis_chem(df, "CBMZ")
# propadiene and NO2
df2 <- df1 <- df
df1$pollutant = "propadiene"
df2$pollutant = "NO2"
(dfe <- rbind(df1, df2))
emis_chem(dfe, "CBMZ")
dfe$region <- rep(letters[1:2], 10)
emis_chem(dfe, "CBMZ", "region")
emis_chem(dfe, "CBMZ", "region", TRUE)
}

#> Warning: number of rows of result is not a multiple of vector length (arg 18)

#>     group region  emission units
#> 2     ALD   <NA> 0.0000000   mol
#> 3  C2H5OH   <NA> 0.0000000   mol
#> 4   CH3OH   <NA> 0.0000000   mol
#> 5      CO   <NA> 0.0000000   mol
#> 6     CSL   <NA> 0.0000000   mol
#> 7     ECC   <NA> 0.0000000     g
#> 8     ECI   <NA> 0.0000000     g
#> 9     ECJ   <NA> 0.0000000     g
#> 10    ETH   <NA> 0.0000000   mol
#> 11    HC3   <NA> 0.0000000   mol
#> 12    HC3      a 0.6240038   mol
#> 13    HC3      b 0.7488045   mol
#> 14    HC5   <NA> 0.0000000   mol
#> 15    HC8   <NA> 0.0000000   mol
#> 16   HCHO   <NA> 0.0000000   mol
#> 17    ISO   <NA> 0.0000000   mol
#> 18    KET   <NA> 0.0000000   mol
#> 19    NH3   <NA> 0.0000000   mol
#> 20     NO   <NA> 0.0000000   mol
#> 21    NO2      a 1.0868148   mol
#> 22    NO2      b 1.3041777   mol
#> 23   NO3C   <NA> 0.0000000     g
#> 24    OL2   <NA> 0.0000000   mol
#> 25   OLE2   <NA> 0.0000000   mol
#> 26    OLI   <NA> 0.0000000   mol
#> 27    OLT   <NA> 0.0000000   mol
#> 28   ORA2   <NA> 0.0000000   mol
#> 29   ORGC   <NA> 0.0000000     g
#> 30   ORGI   <NA> 0.0000000     g
#> 31   ORGJ   <NA> 0.0000000     g
#> 32   PM10   <NA> 0.0000000     g
#> 33  PM25I   <NA> 0.0000000     g
#> 34  PM25J   <NA> 0.0000000     g
#> 35    SO2   <NA> 0.0000000   mol
#> 36   SO4C   <NA> 0.0000000     g
#> 37   SO4I   <NA> 0.0000000     g
#> 38   SO4J   <NA> 0.0000000     g
#> 39    TOL   <NA> 0.0000000   mol
#> 40    XYL   <NA> 0.0000000   mol