Aggregate emissions by lumped groups in chemical mechanism

emis_chem2 aggregates VOC emissions by chemical mechanism and convert grams to mol.

emis_chem2(df, mech, nx, na.rm = FALSE)

Arguments

df
data.frame with emissions including columns "id" and "pol".

mech
Character, "CB4", "CB05", "S99", "S7","CS7", "S7T", "S11", "S11D","S16C","S18B","RADM2", "RACM2","MOZT1"

nx
Character, colnames for emissions data, for instance "V1", "V2"...

na.rm
Logical, to remove lines with NA from group

Value

data.frame with lumped groups by chemical mechanism.

References

Carter, W. P. (2015). Development of a database for chemical mechanism assignments for volatile organic emissions. Journal of the Air & Waste Management Association, 65(10), 1171-1184.

See also

Examples

if (FALSE) {
# experimental
}