Determine deterioration factors for urban conditions

emis_det returns deterioration factors. The emission factors comes from the guidelines for developing emission factors of the EMEP/EEA air pollutant emission inventory guidebook http://www.eea.europa.eu/themes/air/emep-eea-air-pollutant-emission-inventory-guidebook This function subset an internal database of emission factors with each argument

emis_det(po, cc, eu, km)

Arguments

po
Character; Pollutant "CO", "NOx" or "HC"

cc
Character; Size of engine in cc converin "<=1400", "1400_2000" or ">2000"

eu
Character; Euro standard: "I", "II", "III", "III", "IV"

km
Numeric; accumulated mileage in km

Value

It returns a numeric vector representing the increase in emissions due to normal deterioring

Examples

{
data(fkm)
pckm <- fkm[[1]](1:24); pckma <- cumsum(pckm)
(cod1 <- emis_det(po = "CO", cc = "<=1400", eu = "III", km = pckma[1:11]))
}

#>  [1] 1.000000 1.000000 1.038520 1.135721 1.235592 1.337570 1.441121 1.545738
#>  [9] 1.650946 1.756295 1.861367