Merge several emissions files returning data-frames or 'sf' of lines

emis_merge reads rds files and returns a data-frame or an object of 'spatial feature' of streets, merging several files.

emis_merge(
  pol = "CO",
  what = "STREETS.rds",
  streets = T,
  net,
  FN = "sum",
  ignore,
  path = "emi",
  crs,
  under = "after",
  as_list = FALSE,
  k = 1,
  verbose = TRUE
)

Arguments

pol
Character. Pollutant.

what
Character. Word to search the emissions names, "STREETS", "DF" or whatever name. It is important to include the extension .'rds'. For instance, If you have several files "XX_CO_STREETS.rds", what should be "STREETS.rds"

streets
Logical. If true, emis_merge will read the street emissions created with emis_post by "streets_wide", returning an object with class 'sf'. If false, it will read the emissions data-frame and rbind them.

net
'Spatial feature' or 'SpatialLinesDataFrame' with the streets. It is expected #' that the number of rows is equal to the number of rows of street emissions. If #' not, the function will stop.

FN
Character indicating the function. Default is "sum"

ignore
Character; Which pollutants or other charavter would you like to remove?

path
Character. Path where emissions are located

crs
coordinate reference system in numeric format from http://spatialreference.org/ to transform/project spatial data using sf::st_transform

under
"Character"; "after" when you stored your pollutant x as 'X_' "before" when '_X' and "none" for merging directly the files.

as_list
"Logical"; for returning the results as list or not.

k
factor

verbose
Logical to display more information or not. Default is TRUE

Value

'Spatial feature' of lines or a dataframe of emissions

Examples

if (FALSE) {
# Do not run

}