Estimation of exhaust emissions using MOVES

moves_exhaust estimates vehicular emissions as the product of the vehicles on a road, length of the road, emission factor avaliated at the respective speed. \(E = VEH*LENGTH*EF(speed)\)

moves_exhaust(
  veh,
  lkm,
  ef,
  sourceTypeID = 21,
  fuelTypeID = 1,
  pollutantID = 91,
  roadTypeID = 5,
  fuel_type,
  speed_bin,
  profile,
  vehicle = NULL,
  vehicle_type = NULL,
  fuel_subtype = NULL,
  path_all,
  verbose = FALSE
)

Arguments

veh
"Vehicles" data-frame or list of "Vehicles" data-frame. Each data-frame as number of columns matching the age distribution of that ype of vehicle. The number of rows is equal to the number of streets link.

lkm
Length of each link in miles

ef
emission factors from EmissionRates_running exported from MOVES

sourceTypeID
Number to identify type of vehicle as defined by MOVES.

fuelTypeID
Number to identify type of fuel as defined by MOVES.

pollutantID
Number to identify type of pollutant as defined by MOVES.

roadTypeID
Number to identify type of road as defined by MOVES.

fuel_type
Data.frame of fuelSubtypeID exported by MOVES.

speed_bin
Data.frame or vector of avgSpeedBinID as defined by MOVES.

profile
Data.frame or Matrix with nrows equal to 24 and ncol 7 day of the week

vehicle
Character, type of vehicle

vehicle_type
Character, subtype of vehicle

fuel_subtype
Character, subtype of vehicle

path_all
Character to export whole estimation. It is not recommended since it is usually too heavy.

verbose
Logical; To show more information. Not implemented yet

Value

a list with emissions at each street and data.base aggregated by categories. See link{emis_post}

Note

`decoder` shows a decoder for MOVES to identify

Examples

if (FALSE) {
data(decoder)
decoder
}