MOVES estimation of using rates per distance by model year

moves_rpdy_meta estimates running exhaust emissions using MOVES emission factors.

moves_rpdy_meta(
  metadata,
  lkm,
  ef,
  fuel_type,
  speed_bin,
  profile,
  agemax = 31,
  net,
  simplify = TRUE,
  verbose = FALSE
)

Arguments

metadata

data.frame with the metadata for a vein project for MOVES.

lkm

Length of each link in miles

ef

emission factors from EmissionRates_running exported from MOVES

fuel_type

Data.frame of fuelSubtypeID exported by MOVES.

speed_bin

Data.frame or vector of avgSpeedBinID as defined by MOVES.

profile

Data.frame or Matrix with nrows equal to 24 and ncol 7 day of the week

agemax

Integer; max age for the fleet, assuming the same for all vehicles.

net

Road network class sf

simplify

Logical, to return the whole object or processed by streets and veh

verbose

Logical; To show more information. Not implemented yet

Value

a list with emissions at each street and data.base aggregated by categories.

Note

The idea is the user enter with emissions factors by pollutant

Examples

{
data(decoder)
decoder
}
#> $regulatory_class
#>   ID reClassName                                          regClassDesc
#> 1 10          MC                                           Motorcycles
#> 2 20         LDV                                   Light Duty Vehicles
#> 3 30         LDT                                     Light Duty Trucks
#> 4 41      LHD2b3 Class 2b and 3 Trucks (8,500 lbs < GVWR ≤ 14,000 lbs)
#> 5 42       LHD45 Class 4 and 5 Trucks (14,000 lbs < GVWR ≤ 19,500 lbs)
#> 6 46       MHD67 Class 6 and 7 Trucks (19,500 lbs < GVWR ≤ 33,000 lbs)
#> 7 47        HHD8            Class 8a and 8b Trucks (GVWR > 33,000 lbs)
#> 8 48   Urban Bus                      Urban Bus (see CFR Sec 86.091_2)
#> 9 49     Gliders                Glider Vehicles (see EPA-420-F-15-904)
#> 
#> $activity_type
#>    ID              activityType
#> 1   1         Distance traveled
#> 2   2              Source Hours
#> 3   3       Extended Idle Hours
#> 4   4    Source Hours Operating
#> 5   5       Source Hours Parked
#> 6   6                Population
#> 7   7                    Starts
#> 8  13      Hotelling Diesel Aux
#> 9  14   Hotelling Battery or AC
#> 10 15 Hotelling All Engines Off
#> 
#> $emission_process
#>    ID                         processID
#> 1   1                   Running Exhaust
#> 2   2                     Start Exhaust
#> 3   9                         Brakewear
#> 4  10                          Tirewear
#> 5  11                   Evap Permeation
#> 6  12           Evap Fuel Vapor Venting
#> 7  13                   Evap Fuel Leaks
#> 8  15         Crankcase Running Exhaust
#> 9  16           Crankcase Start Exhaust
#> 10 17   Crankcase Extended Idle Exhaust
#> 11 18 Refueling Displacement Vapor Loss
#> 12 19           Refueling Spillage Loss
#> 13 90             Extended Idle Exhaust
#> 14 91           Auxiliary Power Exhaust
#> 
#> $source_type
#>    ID HPMSType               sourceTypeName
#> 1  11       10                   Motorcycle
#> 2  21       25                Passenger Car
#> 3  31       25              Passenger Truck
#> 4  32       25       Light Commercial Truck
#> 5  41       40                  Other Buses
#> 6  42       40                  Transit Bus
#> 7  43       40                   School Bus
#> 8  51       50                 Refuse Truck
#> 9  52       50 Single Unit Short-haul Truck
#> 10 53       50  Single Unit Long-haul Truck
#> 11 54       50                   Motor Home
#> 12 61       60 Combination Short-haul Truck
#> 13 62       60  Combination Long-haul Truck
#> 
#> $fuel_type
#>   ID    fuelType
#> 1  1    Gasoline
#> 2  2      Diesel
#> 3  3         CNG
#> 4  4         LPG
#> 5  5         E85
#> 6  9 Electricity
#> 
#> $road_type
#>   ID                  roadType
#> 1  1               Off-Network
#> 2  2   Rural Restricted Access
#> 3  3 Rural Unrestricted Access
#> 4  4   Urban Restricted Access
#> 5  5 Urban Unrestricted Access
#> 
#> $day_id
#>   ID dayName
#> 1  2 Weekend
#> 2  5 Weekday
#> 
#> $speed_bin
#>    ID lower_speed_limit upper_speed_limit
#> 1   1               0.0               2.5
#> 2   2               2.5               7.5
#> 3   3               7.5              12.5
#> 4   4              12.5              17.5
#> 5   5              17.5              22.5
#> 6   6              22.5              27.5
#> 7   7              27.5              32.5
#> 8   8              32.5              37.5
#> 9   9              37.5              42.5
#> 10 10              42.5              47.5
#> 11 11              47.5              52.5
#> 12 12              52.5              57.5
#> 13 13              57.5              62.5
#> 14 14              62.5              67.5
#> 15 15              67.5              72.5
#> 16 16              72.5              75.0
#> 
#> $running_mode
#>     ID                    name VSP_lower  VSP VSP_higher speed_lower    speed
#> 1    0                 Braking      <NA> <NA>       <NA>        <NA>     <NA>
#> 2    1                  Idling      <NA> <NA>       <NA>        <NA>     <NA>
#> 3   11      Low Speed Coasting      <NA>  VSP        < 0     1 mph ≤    Speed
#> 4   12   Cruise / Acceleration       0 ≤  VSP        < 3     1 mph ≤    Speed
#> 5   13   Cruise / Acceleration       3 ≤  VSP        < 6     1 mph ≤    Speed
#> 6   14   Cruise / Acceleration       6 ≤  VSP        < 9     1 mph ≤    Speed
#> 7   15   Cruise / Acceleration       9 ≤  VSP       < 12     1 mph ≤    Speed
#> 8   16   Cruise / Acceleration      12 ≤  VSP    1 mph ≤       Speed < 25 mph
#> 9   21 Moderate Speed Coasting       VSP  < 0   25 mph ≤       Speed < 50 mph
#> 10  22   Cruise / Acceleration       0 ≤  VSP        < 3    25 mph ≤    Speed
#> 11  23   Cruise / Acceleration       3 ≤  VSP        < 6    25 mph ≤    Speed
#> 12  24   Cruise / Acceleration       6 ≤  VSP        < 9    25 mph ≤    Speed
#> 13  25   Cruise / Acceleration       9 ≤  VSP       < 12    25 mph ≤    Speed
#> 14  27   Cruise / Acceleration      12 ≤  VSP       < 18    25 mph ≤    Speed
#> 15  28   Cruise / Acceleration      18 ≤  VSP       < 24    25 mph ≤    Speed
#> 16  29   Cruise / Acceleration      24 ≤  VSP       < 30    25 mph ≤    Speed
#> 17  30   Cruise / Acceleration      30 ≤  VSP   25 mph ≤       Speed < 50 mph
#> 18  33   Cruise / Acceleration       VSP  < 6   50 mph ≤       Speed     <NA>
#> 19  35   Cruise / Acceleration       6 ≤  VSP       < 12    50 mph ≤    Speed
#> 20  37   Cruise / Acceleration      12 ≤  VSP       < 18    50 mph ≤    Speed
#> 21  38   Cruise / Acceleration      18 ≤  VSP       < 24    50 mph ≤    Speed
#> 22  39   Cruise / Acceleration      24 ≤  VSP       < 30    50 mph ≤    Speed
#> 23  40   Cruise / Acceleration      30 ≤  VSP   50 mph ≤       Speed     <NA>
#> 24 501     Brakewear ; stopped      <NA> <NA>       <NA>        <NA>     <NA>
#>    speed_higher
#> 1          <NA>
#> 2          <NA>
#> 3      < 25 mph
#> 4      < 25 mph
#> 5      < 25 mph
#> 6      < 25 mph
#> 7      < 25 mph
#> 8          <NA>
#> 9          <NA>
#> 10     < 50 mph
#> 11     < 50 mph
#> 12     < 50 mph
#> 13     < 50 mph
#> 14     < 50 mph
#> 15     < 50 mph
#> 16     < 50 mph
#> 17         <NA>
#> 18         <NA>
#> 19         <NA>
#> 20         <NA>
#> 21         <NA>
#> 22         <NA>
#> 23         <NA>
#> 24         <NA>
#> 
#> $starts_mode
#>    ID   soak_time_i soak_time   soak_time_j
#> 1 101          <NA> Soak Time   < 6 minutes
#> 2 102   6 minutes ≤ Soak Time  < 30 minutes
#> 3 103  30 minutes ≤ Soak Time  < 60 minutes
#> 4 104  60 minutes ≤ Soak Time  < 90 minutes
#> 5 105  90 minutes ≤ Soak Time < 120 minutes
#> 6 106 120 minutes ≤ Soak Time < 360 minutes
#> 7 107 360 minutes ≤ Soak Time < 720 minutes
#> 8 108 720 minutes ≤ Soak Time          <NA>
#> 
#> $hotelling_mode
#>    ID                 hotelling
#> 1 200           Extended Idling
#> 2 201 Auxiliary Power Units Use
#> 3 202             Battery Power
#> 4 204                Engine off
#> 
#> $pollutants
#>      ID                                 pollutant short_name
#> 1     1                             pollutantname        THC
#> 2     2                           Carbon Monoxide         CO
#> 3     3                        Oxides of Nitrogen        NOx
#> 4     5                                   Methane        CH4
#> 5     6                             Nitrous Oxide        N2O
#> 6    20                                   Benzene        BEN
#> 7    21                                   Ethanol       ETHO
#> 8    23                      Naphthalene particle       NAPH
#> 9    24                             1,3-Butadiene     1,3BUT
#> 10   25                              Formaldehyde       HCHO
#> 11   26                              Acetaldehyde        ALD
#> 12   27                                  Acrolein       ACRO
#> 13   30                                   Ammonia        NH3
#> 14   31                            Sulfur Dioxide        SO2
#> 15   32                            Nitrogen Oxide         NO
#> 16   33                          Nitrogen Dioxide        NO2
#> 17   34                              Nitrous Acid       HONO
#> 18   35                                   Nitrate        NO3
#> 19   36                                  Ammonium        NH4
#> 20   40                    2,2,4-Trimethylpentane       <NA>
#> 21   41                             Ethyl Benzene       <NA>
#> 22   42                                    Hexane       <NA>
#> 23   43                           Propionaldehyde       <NA>
#> 24   44                                   Styrene       <NA>
#> 25   45                                   Toluene       <NA>
#> 26   46                                    Xylene       <NA>
#> 27   51                                  Chloride       <NA>
#> 28   52                                    Sodium       <NA>
#> 29   53                                 Potassium       <NA>
#> 30   54                                 Magnesium       <NA>
#> 31   55                                   Calcium       <NA>
#> 32   56                                  Titanium       <NA>
#> 33   57                                   Silicon       <NA>
#> 34   58                                  Aluminum       <NA>
#> 35   59                                      Iron       <NA>
#> 36   60                 Mercury Elemental Gaseous       <NA>
#> 37   61                  Mercury Divalent Gaseous       <NA>
#> 38   62                       Mercury Particulate       <NA>
#> 39   63                         Arsenic Compounds       <NA>
#> 40   65                               Chromium 6+       <NA>
#> 41   66                       Manganese Compounds       <NA>
#> 42   67                          Nickel Compounds       <NA>
#> 43   68           Dibenzo(a,h)anthracene particle       <NA>
#> 44   69                     Fluoranthene particle       <NA>
#> 45   70                     Acenaphthene particle       <NA>
#> 46   71                   Acenaphthylene particle       <NA>
#> 47   72                       Anthracene particle       <NA>
#> 48   73                Benz(a)anthracene particle       <NA>
#> 49   74                   Benzo(a)pyrene particle       <NA>
#> 50   75             Benzo(b)fluoranthene particle       <NA>
#> 51   76             Benzo(g,h,i)perylene particle       <NA>
#> 52   77             Benzo(k)fluoranthene particle       <NA>
#> 53   78                         Chrysene particle       <NA>
#> 54   79                  Non-Methane Hydrocarbons       NMHC
#> 55   80                 Non-Methane Organic Gases       NMOG
#> 56   81                         Fluorene particle       <NA>
#> 57   82          Indeno(1,2,3,c,d)pyrene particle       <NA>
#> 58   83                     Phenanthrene particle       <NA>
#> 59   84                           Pyrene particle       <NA>
#> 60   86                       Total Organic Gases        TOG
#> 61   87                Volatile Organic Compounds        VOC
#> 62   88                                 NonHAPTOG       <NA>
#> 63   90                           Atmospheric CO2        CO2
#> 64   91                  Total Energy Consumption        TEC
#> 65   92              Petroleum Energy Consumption        PEC
#> 66   93            Fossil Fuel Energy Consumption       FFEC
#> 67   98                            CO2 Equivalent       CO2E
#> 68   99    Brake Specific Fuel Consumption (BSFC)       <NA>
#> 69  100              Primary Exhaust PM10 - Total       PM10
#> 70  106      Primary PM10 - Brakewear Particulate     PM10BW
#> 71  107       Primary PM10 - Tirewear Particulate     PM10TW
#> 72  110             Primary Exhaust PM2.5 - Total      PM2.5
#> 73  111                            Organic Carbon         OC
#> 74  112                          Elemental Carbon         EC
#> 75  115                       Sulfate Particulate       <NA>
#> 76  116     Primary PM2.5 - Brakewear Particulate    PM2.5BW
#> 77  117      Primary PM2.5 - Tirewear Particulate    PM2.5TW
#> 78  118                       Composite - NonECPM       <NA>
#> 79  119                             H2O (aerosol)       <NA>
#> 80  120             Primary PM2.5 - NonECNonSO4PM       <NA>
#> 81  121              CMAQ5.0 Unspeciated (PMOTHR)       <NA>
#> 82  122          Non-carbon Organic Matter (NCOM)       <NA>
#> 83  130    1,2,3,7,8,9-Hexachlorodibenzo-p-Dioxin       <NA>
#> 84  131                Octachlorodibenzo-p-dioxin       <NA>
#> 85  132 1,2,3,4,6,7,8-Heptachlorodibenzo-p-Dioxin       <NA>
#> 86  133                    Octachlorodibenzofuran       <NA>
#> 87  134    1,2,3,4,7,8-Hexachlorodibenzo-p-Dioxin       <NA>
#> 88  135     1,2,3,7,8-Pentachlorodibenzo-p-Dioxin       <NA>
#> 89  136           2,3,7,8-Tetrachlorodibenzofuran       <NA>
#> 90  137     1,2,3,4,7,8,9-Heptachlorodibenzofuran       <NA>
#> 91  138         2,3,4,7,8-Pentachlorodibenzofuran       <NA>
#> 92  139         1,2,3,7,8-Pentachlorodibenzofuran       <NA>
#> 93  140        1,2,3,6,7,8-Hexachlorodibenzofuran       <NA>
#> 94  141    1,2,3,6,7,8-Hexachlorodibenzo-p-Dioxin       <NA>
#> 95  142       2,3,7,8-Tetrachlorodibenzo-p-Dioxin       <NA>
#> 96  143        2,3,4,6,7,8-Hexachlorodibenzofuran       <NA>
#> 97  144     1,2,3,4,6,7,8-Heptachlorodibenzofuran       <NA>
#> 98  145        1,2,3,4,7,8-Hexachlorodibenzofuran       <NA>
#> 99  146        1,2,3,7,8,9-Hexachlorodibenzofuran       <NA>
#> 100 168                Dibenzo(a,h)anthracene gas       <NA>
#> 101 169                          Fluoranthene gas       <NA>
#> 102 170                          Acenaphthene gas       <NA>
#> 103 171                        Acenaphthylene gas       <NA>
#> 104 172                            Anthracene gas       <NA>
#> 105 173                     Benz(a)anthracene gas       <NA>
#> 106 174                        Benzo(a)pyrene gas       <NA>
#> 107 175                  Benzo(b)fluoranthene gas       <NA>
#> 108 176                  Benzo(g,h,i)perylene gas       <NA>
#> 109 177                  Benzo(k)fluoranthene gas       <NA>
#> 110 178                              Chrysene gas       <NA>
#> 111 181                              Fluorene gas       <NA>
#> 112 182               Indeno(1,2,3,c,d)pyrene gas       <NA>
#> 113 183                          Phenanthrene gas       <NA>
#> 114 184                                Pyrene gas       <NA>
#> 115 185                           Naphthalene gas       <NA>
#>