MOVES estimation of using rates per distance by model year

moves_rpdy_sf estimates running exhaust emissions using MOVES emission factors.

moves_rpdy_sf(
  veh,
  lkm,
  ef,
  speed_bin,
  profile,
  source_type_id = 21,
  vehicle = NULL,
  vehicle_type = NULL,
  fuel_subtype = NULL,
  net,
  path_all,
  verbose = FALSE
)

Arguments

veh
"Vehicles" data-frame or list of "Vehicles" data-frame. Each data-frame as number of columns matching the age distribution of that ype of vehicle. The number of rows is equal to the number of streets link.

lkm
Length of each link in miles

ef
emission factors from EmissionRates_running exported from MOVES filtered by sourceTypeID and fuelTypeID.

speed_bin
Data.frame or vector of avgSpeedBinID as defined by MOVES.

profile
numeric vector of normalized traffic for the morning rush hour

source_type_id
Number to identify type of vehicle as defined by MOVES.

vehicle
Character, type of vehicle

vehicle_type
Character, subtype of vehicle

fuel_subtype
Character, subtype of vehicle

net
Road network class sf

path_all
Character to export whole estimation. It is not recommended since it is usually too heavy.

verbose
Logical; To show more information. Not implemented yet

Value

a list with emissions at each street and data.base aggregated by categories. See link{emis_post}

Note

`decoder` shows a decoder for MOVES

Examples

{
data(decoder)
decoder
}

#>    CategoryField pollutantID                            Description
#> 1          dayID           5                                Weekday
#> 2          dayID           2                                Weekend
#> 3   sourceTypeID          11                             Motorcycle
#> 4   sourceTypeID          21                          Passenger Car
#> 5   sourceTypeID          31                        Passenger Truck
#> 6   sourceTypeID          32                 Light Commercial Truck
#> 7   sourceTypeID          41                            Other Buses
#> 8   sourceTypeID          42                            Transit Bus
#> 9   sourceTypeID          43                             School Bus
#> 10  sourceTypeID          51                           Refuse Truck
#> 11  sourceTypeID          52           Single Unit Short-haul Truck
#> 12  sourceTypeID          53            Single Unit Long-haul Truck
#> 13  sourceTypeID          54                             Motor Home
#> 14  sourceTypeID          61           Combination Short-haul Truck
#> 15  sourceTypeID          62            Combination Long-haul Truck
#> 16    roadTypeID           1                            Off-Network
#> 17    roadTypeID           2                Rural Restricted Access
#> 18    roadTypeID           3              Rural Unrestricted Access
#> 19    roadTypeID           4                Urban Restricted Access
#> 20    roadTypeID           5              Urban Unrestricted Access
#> 21    fuelTypeID           1                               Gasoline
#> 22    fuelTypeID           2                            Diesel Fuel
#> 23    fuelTypeID           3           Compressed Natural Gas (CNG)
#> 24    fuelTypeID           5                         Ethanol (E-85)
#> 25   pollutantID           1             Total Gaseous Hydrocarbons
#> 26   pollutantID           2                        Carbon Monoxide
#> 27   pollutantID           3                     Oxides of Nitrogen
#> 28   pollutantID           5                                Methane
#> 29   pollutantID           6                          Nitrous Oxide
#> 30   pollutantID          20                                Benzene
#> 31   pollutantID          21                                Ethanol
#> 32   pollutantID          30                                Ammonia
#> 33   pollutantID          31                        Sulfure Dioxide
#> 34   pollutantID          32                         Nitrogen Oxide
#> 35   pollutantID          33                       Nitrogen Dioxide
#> 36   pollutantID          34                           Nitrous Acid
#> 37   pollutantID          79               Non-Methane Hydrocarbons
#> 38   pollutantID          86                    Total Organic Gases
#> 39   pollutantID          87             Volatile Organic Compounds
#> 40   pollutantID          90                        Atmospheric CO2
#> 41   pollutantID          91               Total Energy Consumption
#> 42   pollutantID          92           Petroleum Energy Consumption
#> 43   pollutantID          93         Fossil Fuel Energy Consumption
#> 44   pollutantID          98                         CO2 Equivalent
#> 45   pollutantID         100           Primary Exhaust PM10 - Total
#> 46   pollutantID         101          Primary PM10 - Organic Carbon
#> 47   pollutantID         102        Primary PM10 - Elemental Carbon
#> 48   pollutantID         105     Primary PM10 - Sulfate Particulate
#> 49   pollutantID         106  Primary PM10 -  Brakewear Particulate
#> 50   pollutantID         107   Primary PM10 -  Tirewear Particulate
#> 51   pollutantID         110          Primary Exhaust PM2.5 - Total
#> 52   pollutantID         111         Primary PM2.5 - Organic Carbon
#> 53   pollutantID         112       Primary PM2.5 - Elemental Carbon
#> 54   pollutantID         115    Primary PM2.5 - Sulfate Particulate
#> 55   pollutantID         116 Primary PM2.5 -  Brakewear Particulate
#> 56   pollutantID         117  Primary PM2.5 -  Tirewear Particulate
#> 57     processID           1                        Running Exhaust
#> 58     processID           2                          Start Exhaust
#> 59     processID          15                      Crank Run Exhaust
#> 60     processID          16                    Crank Start Exhaust
#> 61     processID          17                 Crank Ext Idle Exhaust
#> 62     processID          90                       Ext Idle Exhaust
#> 63     processID          11                        Evap Permeation
#> 64     processID          12                  Evaporative Fuel Vent
#> 65     processID          13                  Evaporative Fuel Leak
#> 66     processID          18                        Refuel Disp Vap
#> 67     processID          19                        Refuel Spillage
#> 68     processID           9                              Brakewear
#> 69     processID          10                               Tirewear
#>              V4
#> 1              
#> 2              
#> 3              
#> 4              
#> 5              
#> 6              
#> 7              
#> 8              
#> 9              
#> 10             
#> 11             
#> 12             
#> 13             
#> 14             
#> 15             
#> 16             
#> 17             
#> 18             
#> 19             
#> 20             
#> 21             
#> 22             
#> 23             
#> 24             
#> 25           HC
#> 26           CO
#> 27          NOx
#> 28          CH4
#> 29          N2O
#> 30      BENZENE
#> 31         ETOH
#> 32          NH3
#> 33          SO2
#> 34          NOx
#> 35          NO2
#> 36         HONO
#> 37         NMHC
#> 38          TOG
#> 39          VOC
#> 40       AT CO2
#> 41    TotEnergy
#> 42    PetEnergy
#> 43 FossilEnergy
#> 44        CO2EQ
#> 45   PM10TotExh
#> 46       PM10OC
#> 47       PM10EC
#> 48  PM10Sulfate
#> 49    PM10Brake
#> 50     PM10Tire
#> 51  PM2.5TotExh
#> 52      PM2.5OC
#> 53      PM2.5EC
#> 54 PM2.5Sulfate
#> 55   PM2.5Brake
#> 56    PM2.5Tire
#> 57             
#> 58             
#> 59             
#> 60             
#> 61             
#> 62             
#> 63             
#> 64             
#> 65             
#> 66             
#> 67             
#> 68             
#> 69