Calculate speeds of traffic network

Creates a dataframe of speeds fir diferent hours and each link based on morning rush traffic data

netspeed(q, ps, ffs, cap, lkm, alpha = 0.15, beta = 4, isList = FALSE,
  distance = "km", time = "h")

Arguments

q
Data-frame of traffic flow to each hour (veh/h)

ps
Peak speed (km/h)

ffs
Free flow speed (km/h)

cap
Capacity of link (veh/h)

lkm
Distance of link (km)

alpha
Parameter of BPR curves

beta
Parameter of BPR curves

isList
Logical to specify type of return, list or data-frame

distance
Character specifying the units for distance. Default is "km"

time
Character specifying the units for time Default is "h"

Value

dataframe or list of speeds with units

Examples

# NOT RUN {
# Do not run
data(net)
data(pc_profile)
pc_week <- temp_fact(net$ldv+net$hdv, pc_profile)
df <- netspeed(pc_week, net$ps, net$ffs, net$capacity, net$lkm)
class(df)
plot(df) #plot of the average speed at each hour, +- sd
# }