Estimation of average daily running losses evaporative emissions

running_losses estimates evaporative emissions from EMEP/EEA emisison guidelines

running_losses(x, carb, p, erhotc, erwarmc, erhotfi)

Arguments

x
Mean number of trips per vehicle per day

carb
fraction of gasoline vehicles with carburator or fuel return system

p
Fraction of trips finished with hot engine

erhotc
average daily running losses evaporative factor for vehicles with carburator or fuel return system

erwarmc
average daily cold and warm running losses evaporative factor for vehicles with carburator or fuel return system

erhotfi
average daily hot running losses evaporative factor for vehicles with fuel injection and returnless fuel systems

Value

numeric vector of emission estimation in grams

References

Mellios G and Ntziachristos 2016. Gasoline evaporation. In: EEA, EMEP. EEA air pollutant emission inventory guidebook-2009. European Environment Agency, Copenhagen, 2009

Examples

{
# Do not run
ev <- running_losses(x = 1:10, carb = 0, p = 1, erhot = 1, erwarmc =1,
erhotfi = 1)
summary(ev)
}

#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>    1.00    3.25    5.50    5.50    7.75   10.00