Speciation of emissions

The separation of emissions in different compunds. It includes black carbon and organic matter from particulate matter. Soon it will be added more speciations

speciate(x, spec = "bcom", veh, fuel, eu, show = FALSE, list = FALSE)

Arguments

x
Emissions estimation

spec
type of speciation, e.g.: "bcom" stands for black carbon and organic matter. The speciations are: "bcom", tyre", "break", "road","iag" and "nox". 'iag' now includes a speciation for use of industrial and building paintings

veh
Type of vehicle. When spec is "bcom" or "nox" veh can be "PC", "LCV", HDV" or "Motorcycle". When spec is "iag" veh can take two values depending: when the speciation is for vehicles veh accepts "veh", eu "Evaporative", "Liquid" or "Exhaust" and fuel "G", "E" or "D", when the speciation is for painting, veh is "paint" fuel or eu can be "industrial" or "buildind".

fuel
Fuel. When spec is "bcom" fuel can be "G" or "D". When spec is "iag" fuel can be "G", "E" or "D". When spec is "nox" fuel can be "G", "D", "LPG", "E85" or "CNG". Not required for "tyre", "break" or "road"

eu
Euro emission standard: "PRE", "ECE_1501", "ECE_1502", "ECE_1503", "I", "II", "III", "IV", "V", "III-CDFP","IV-CDFP","V-CDFP", "III-ADFP", "IV-ADFP","V-ADFP" and "OPEN_LOOP". When spec is "iag" accept the values "Exhaust" "Evaporative" and "Liquid". When spec is "nox" eu can be "PRE", "I", "II", "III", "IV", "V", "VI", "VIc", "III-DPF" or "III+CRT". Not required for "tyre", "break" or "road"

show
when TRUE shows row of table with respective speciation

list
when TRUE returns a list with number of elements of the list as the number species of pollutants

Value

dataframe of speciation in grams or mols

Note

when spec = "iag", veh is only "VEH", STANDARD is "Evaporative", "Liquid" or "Exhaust", FUEL is "G" for gasoline (blended with 25% ethanol), "E" for Ethanol and "D" for diesel (blended with 5% of biodiesel). When spec = "bcom", veh can be "PC", "LCV", "Motorcycle" or "HDV" VEH", STANDARD is "Evaporative", "Liquid" or "Exhaust", FUEL is "G" for gasoline (blended with 25% ethanol), "E" for Ethanol and "D" for diesel (blended with 5% of biodiesel).

References

"bcom": Ntziachristos and Zamaras. 2016. Passneger cars, light commercial trucks, heavy-duty vehicles including buses and motor cycles. In: EEA, EMEP. EEA air pollutant emission inventory guidebook-2009. European Environment Agency, Copenhagen, 2016

"tyre", "break" and "road": Ntziachristos and Boulter 2016. Automobile tyre and break wear and road abrasion. In: EEA, EMEP. EEA air pollutant emission inventory guidebook-2009. European Environment Agency, Copenhagen, 2016

"iag": RAFEE, S.A.A. Estudo numerico do impacto das emissoes veiculares e fixas da cidade de Manaus nas concentracoes de poluentes atmosfericos da regiao amazonica. 2015. 109 f. Dissertacao (Mestrado). Programa de Pos-Graduacao em Engenharia Ambiental (PPGEA) - Universidade Tecnologica Federal do Parana. Londrina, 2015. http://repositorio.utfpr.edu.br/jspui/bitstream/1/1675/1/LD_PPGEA_M_Rafee

"iag": Vela, A. L. V. Avaliacao do impacto da mudanca dos fatores de emissao veicular na formacao de ozonio troposferico na Regiao Metropolitana de Sao Paulo. 2013. Dissertacao de Mestrado. Instituto de Astronomia, Geofisica e Ciencias Atmosfericas, Universidade de Sao Paulo, Sao Paulo. http://www.iag.usp.br/pos/sites/default/files/d_angel_l_v_vela_corrigida_0.pdf Speciate EPA: https://cfpub.epa.gov/speciate/. : K. Sexton, H. Westberg, "Ambient hydrocarbon and ozone measurements downwind of a large automotive painting plant" Environ. Sci. Tchnol. 14:329 (1980).P.A. Scheff, R.A. Schauer, James J., Kleeman, Mike J., Cass, Glen R., Characterization and Control of Organic Compounds Emitted from Air Pollution Sources, Final Report, Contract 93-329, prepared for California Air Resources Board Research Division, Sacramento, CA, April 1998. 2004 NPRI National Databases as of April 25, 2006, http://www.ec.gc.ca/pdb/npri/npri_dat_rep_e.cfm. Memorandum Proposed procedures for preparing composite speciation profiles using Environment Canada s National Pollutant Release Inventory (NPRI) for stationary sources, prepared by Ying Hsu and Randy Strait of E.H. Pechan Associates, Inc. for David Niemi, Marc Deslauriers, and Lisa Graham of Environment Canada, September 26, 2006.

Examples

# NOT RUN {
# Do not run
pm <- rnorm(n = 100, mean = 400, sd = 2)
df <- speciate(pm, veh="PC", fuel="G", eu="I")
# }