creates a .tex a table from a data.frame

to_latex reads a data.frme an dgenerates a .tex table, aiming to replicate the method of tablegenerator.com

to_latex(df, file, caption = "My table", label = "tab:df")

Arguments

df
data.frame with three column.

file
Character, name of new .tex file

caption
Character caption of table

label
Character, label of table

Value

a text file with extension .tex.

See also

Examples

{
df <- data.frame(pollutant = rep(c("CO", "propadiene", "NO2"), 10),
                 emission = vein::Emissions(1:30),
                 region = rep(letters[1:2], 15))
df
long_to_wide(df)
(df2 <- long_to_wide(df, column_fixed = "region"))
to_latex(df2)
to_latex(long_to_wide(df, column_fixed = "region"),
file = paste0(tempfile(), ".tex"))
}

#> % Generated with vein 0.8.2
#> % Please add the following required packages to your document preamble:
#> % \usepackage{booktabs}
#> % \usepackage{graphicx}
#> \begin{table}[]
#> \caption{My table}
#> \label{tab:df}
#> \resizebox{\textwidth}{!}{%
#> \begin{tabular}{@{} l r r r @{}}
#> \toprule
#> region &  CO &  NO2 &  propadiene \\ \midrule
#> a &  1 &  3 &  5 \\  
#> a &  7 &  9 &  11 \\  
#> a &  13 &  15 &  17 \\  
#> a &  19 &  21 &  23 \\  
#> a &  25 &  27 &  29 \\  
#> b &  4 &  6 &  2 \\  
#> b &  10 &  12 &  8 \\  
#> b &  16 &  18 &  14 \\  
#> b &  22 &  24 &  20 \\  
#> b &  28 &  30 &  26 \\  \bottomrule 
#> \end{tabular}%
#> }
#> \end{table}
#> /tmp/RtmpB8y1ve/file6c2d37a62ede.tex