vein: a package for elaborating vehicular emissions inventories

This package provides functions to arrange traffic data, prepare emission factors, estimate emissions and process emissions

Details

1) Inventory

It is recommended to start with the function inventory which produces a set of directories and scripts to run vein.

2) Traffic data

The user must count with traffic data at each street at least for one hour. The format of the data must be spatial, either "SpatialLinesDataFrame" or an object class of "sf". Then the user must use any ob tje age functions: age_ldv, age_hdv, age_moto or my_age. The outputs of these functions can be saved in directory 'veh' with the extension .rds.

3) Emission factors

The user must chosse a type of emission factor: from Copert with the ef_ldv_speed or ef_hdv_speed, from local sources as one constant emission factors by age of use of vehicles with EmissionFactorsList or as a merge between both with ef_ldv_scaled or ef_hdv_scaled.

4) Estimating emissions

Once all information is obtained, the user can estimate the emissions with emis, emis_cold or other.

5) Processing the emissions

The function for processing the emissions ins emis_post.