Cold-Start Emissions factors for Light Duty Vehicles

This function returns speed functions which depends on ambient temperature average speed. The emission factors comes from the guidelines EMEP/EEA air pollutant emission inventory guidebook http://www.eea.europa.eu/themes/air/emep-eea-air-pollutant-emission-inventory-guidebook

ef_ldv_cold(v = "LDV", ta, cc, f, eu, p, k = 1, show.equation = FALSE)

Arguments

v
Category vehicle: "LDV"

ta
Ambient temperature. Monthly men can be used

cc
Size of engine in cc: "<=1400", "1400_2000" or ">2000"

f
Type of fuel: "G", "D" or "LPG"

eu
Euro standard: "PRE", "I", "II", "III", "IV", "V", "VI" or "VIc"

p
Pollutant: "CO", "FC", "NOx", "HC" or "PM"

k
Multiplication factor

show.equation
Option to see or not the equation parameters

Value

an emission factor function which depends of the average speed V and ambient temperature. g/km

Examples

{
ef1 <- ef_ldv_cold(ta = 15, cc = "<=1400", f ="G", eu = "I", p = "CO")
ef1(10)
}

#> [1] 2.754