List of cold start emission factors of Light Duty Vehicles

This function creates a list of functions of cold start emission factors considering different euro emission standard to the elements of the list.

ef_ldv_cold_list(df, v = "LDV", ta, cc, f, eu, p)

Arguments

df
Dataframe with local emission factor

v
Category vehicle: "LDV"

ta
ambient temperature. Montly average van be used

cc
Size of engine in cc: <=1400", "1400_2000" and ">2000"

f
Type of fuel: "G" or "D"

eu
character vector of euro standards: "PRE", "I", "II", "III", "IV", "V", "VI" or "VIc".

p
Pollutant: "CO", "FC", "NOx", "HC" or "PM"

Value

A list of cold start emission factors g/km

Note

The length of the list should be equal to the name of the age categories of a specific type of vehicle

Examples

{
# Do not run
df <- data.frame(age1 = c(1,1), age2 = c(2,2))
eu = c("I", "PRE")
l <- ef_ldv_cold(t = 17, cc = "<=1400", f = "G",
eu = "I", p = "CO")
l_cold <- ef_ldv_cold_list(df, t = 17, cc = "<=1400", f = "G",
eu = eu, p = "CO")
length(l_cold)
}

#> [1] 2