Emissions factors of N2O and NH3

ef_nitro returns emission factors as a functions of accumulated mileage. The emission factors comes from the guidelines EMEP/EEA air pollutant emission inventory guidebook http://www.eea.europa.eu/themes/air/emep-eea-air-pollutant-emission-inventory-guidebook

ef_nitro(v, t, cc, f, eu, p, S, k = 1, show.equation = TRUE)

Arguments

v
Category vehicle: "PC", "LCV", "LDV", "Motorcycle", "Trucks", "HDV", "HDV-A", "BUS" or "Coach".

t
Type: "Cold", "Hot", "<50", ">=50", ">3.5", "7.5_12", "12_18", "28_34", ">34" and "ALL".

cc
"Urban", "Rural", "Highway" and "ALL".

f
Type of fuel: "G", "D" or "LPG"

eu
Euro standard: "PRE", "I", "II", "III", "IV", "V", "VI", "VIc", "2S", 4S" and "ALL"

p
Pollutant: "N2O", "NH3"

S
Sulphur (ppm). Number.

k
Multiplication factor

show.equation
Option to see or not the equation parameters

Value

an emission factor function which depends on the accumulated mileage

Examples

{
# Do not run
efe10 <- ef_nitro(v = "PC", t = "Hot", cc = "Urban", f = "G",
eu = "III", p = "NH3", S = 10,
show.equation = FALSE)
efe50 <- ef_nitro(v = "PC", t = "Hot", cc = "Urban", f = "G",
eu = "III", p = "NH3", S = 50,
show.equation = TRUE)
efe10(10)
efe50(10)
}

#> [[1]]
#> [1] 1.31e-06
#> 
#> [[2]]
#> [1] 0.862
#> 
#> [[3]]
#> [1] 0.0019
#> 
#> [[4]]
#> [1] 4.18e-06
#> 
#> [[5]]
#> [1] 0.526
#> 
#> [[6]]
#> [1] 0.0016
#> 
#> [[7]]
#> [1] 0
#> 
#> [[8]]
#> [1] 0
#> 
#> [[9]]
#> [1] 0
#> 
#> [[10]]
#> [1] "(ifelse(S<=30,(a*km+b)*ab,(c*km+d)*cd))*1"
#> 

#> [1] 0.0008416669