Estimation of cold start emissions with top-down approach

emis_cold_td estimates cld start emissions with a top-down appraoch. This is, annual or monthly emissions or region. Especifically, the emissions are esitmated for row of the simple feature (row of the spatial feature).

In general was designed so that each simple feature is a region with different average monthly temperature. This funcion, as other in this package, adapts to the class of the input data. providing flexibility to the user.

emis_cold_td(veh, lkm, ef, efcold, beta, pro_month, verbose = FALSE)

Arguments

veh
"Vehicles" data-frame or spatial feature, wwhere columns are the age distribution of that vehicle. and rows each simple feature or region. The number of rows is equal to the number of streets link

lkm
Numeric; mileage by the age of use of each vehicle.

ef
Numeric; emission factor with

efcold
Data.frame or list of data.frames. When it is a data.frame, each column is for each type of vehicle by age of use, rows are are each simple feature. When it is a list of data.frames, each element of the list it is a month of the year. efcold should be different for each simple feature, assuming different average temperature.

beta
Data.frame with the fraction of cold starts. The rows are the fraction for each spatial feature or subregion, the columns are the age of use of vehicle.

pro_month
Numeric; montly profile to distribuite annual mileage in each month.

verbose
Logical; To show more information

Value

Emissions data.frame

Examples

# NOT RUN {
# Do not run
# }