Estimation of resuspension emissions from paved roads

emis_paved estimates vehicular emissions from paved roads. The vehicular emissions are estimated as the product of the vehicles on a road, length of the road, emission factor from AP42 13.2.1 Paved roads. It is assumed dry hours and anual aggregation should consider moisture factor. It depends on Average Daily Traffic (ADT)

emis_paved(
  veh,
  adt,
  lkm,
  k = 0.62,
  sL1 = 0.6,
  sL2 = 0.2,
  sL3 = 0.06,
  sL4 = 0.03,
  W,
  net = net
)

Arguments

veh

Numeric vector with length of elements equals to number of streets It is an array with dimenssions number of streets x hours of day x days of week

adt

Numeric vector of with Average Daily Traffic (ADT)

lkm

Length of each link

k

K_PM30 = 3.23 (g/vkm), K_PM15 = 0.77 (g/vkm), K_PM10 = 0.62 (g/vkm) and K_PM2.5 = 0.15 (g/vkm).

sL1

Silt loading (g/m2) for roads with ADT <= 500

sL2

Silt loading (g/m2) for roads with ADT > 500 and <= 5000

sL3

Silt loading (g/m2) for roads with ADT > 5000 and <= 1000

sL4

Silt loading (g/m2) for roads with ADT > 10000

W

array of dimensions of veh. It consists in the hourly averaged weight of traffic fleet in each road

net

SpatialLinesDataFrame or Spatial Feature of "LINESTRING"

Value

emission estimation g/h

Note

silt values can vary a lot. For comparison:

ADT	US-EPA g/m2	CENMA (Chile) g/m2

< 500	0.6	2.4

500-5000	0.2	0.7

5000-1000	0.06	0.6

>10000	0.03	0.3

References

EPA, 2016. Emission factor documentation for AP-42. Section 13.2.1, Paved Roads. https://www3.epa.gov/ttn/chief/ap42/ch13/final/c13s0201.pdf

CENMA Chile: Actualizacion de inventario de emisiones de contaminntes atmosfericos RM 2020 Universidad de Chile#'

Examples

if (FALSE) {
# Do not run
veh <- matrix(1000, nrow = 10,ncol = 10)
W <- veh*1.5
lkm <-  1:10
ADT <-1000:1010
emi  <- emis_paved(veh = veh, adt = ADT, lkm = lkm, k = 0.65, W = W)
class(emi)
head(emi)
}