Speciation of emissions

speciate separates emissions in different compounds. It covers black carbon and organic matter from particulate matter. Soon it will be added more speciations

speciate(x, spec = "bcom", veh, fuel, eu, list = FALSE, pmpar, verbose = FALSE)

Arguments

x
Emissions estimation

spec
The speciations are:

"bcom": Splits PM2.5 in black carbon and organic matter.

"tyre" or "tire": Splits PM in PM10, PM2.5, PM1 and PM0.1.

"brake": Splits PM in PM10, PM2.5, PM1 and PM0.1.

"road": Splits PM in PM10 and PM2.5.

"nox": Splits NOx in NO and NO2.

"nmhc": Splits NMHC in compounds, see ef_ldv_speed.

"pmiag", "pmneu", "pmneu2": Splits PM in groups, see note below.

The following still available but they will be removed soon:

"iag_racm": ethanol emissions added in hc3.

"iag" or "iag_cb05": Splits NMHC by CB05 (WRF exb05_opt1) group .

"petroiag_cb05": Splits NMHC by CB05 (WRF exb05_opt1) group .

"iag_cb05v2": Splits NMHC by CB05 (WRF exb05_opt2) group .

"neu_cb05": Splits NMHC by CB05 (WRF exb05_opt2) group alternative.

"petroiag_cb05v2": Splits NMHC by CB05 (WRF exb05_opt2) group alternative.

veh
Type of vehicle:

"bcom": veh can be "PC", "LCV", HDV" or "Motorcycle".

"tyre" or "tire": not necessary.

"brake": not necessary.

"road": not necessary.

"nox": veh can be "PC", "LCV", HDV" or "Motorcycle".

"nmhc": veh can be "LDV", "HDV" or "LPG".

""pmiag", "pmneu", "pmneu2": not necessary.

#' The following still available but they will be removed soon:

"iag_racm": veh accepts "veh".

"iag" or "iag_cb05": veh accepts "veh" .

"iag_cb05v2": veh accepts "veh" .

"neu_cb05": veh accepts "veh" .

"petroiag_cb05": veh accepts "veh" .

"petroiag_cb05v2": veh accepts "veh" .

fuel
Fuel.

"bcom": "G" or "D".

"tyre" or "tire": not necessary.

"brake": not necessary.

"road": not necessary.

"nox": "G", "D", "LPG", "E85" or "CNG".

"nmhc": "G", "D" or "LPG".

"pmiag", "pmneu", "pmneu2": not necessary.

#' The following still available but they will be removed soon:

"iag_racm": "G", "E" or "D".

"iag" or "iag_cb05": "G", "E" or "D".

"iag_cb05v2": "G", "E" or "D".

"neu_cb05": "G", "E" or "D".

"petroiag_cb05": "G", "E" or "D".

"petroiag_cb05v2": "G", "E" or "D".

eu
Emission standard

"bcom": "G" or "D".

"tyre" or "tire": not necessary.

"brake": not necessary.

"road": not necessary.

"nox": "G", "D", "LPG", "E85" or "CNG".

"nmhc": "PRE", "ECE_1501", "ECE_1502", "ECE_1503","I", "II", "III", "IV", "V", "III-CDFP","IV-CDFP","V-CDFP", "III-ADFP", "IV-ADFP","V-ADFP" or "OPEN_LOOP". eu can be "ALL" if spec is "nmhc" and fuel is "LPG"

"pmiag", "pmneu", "pmneu2": not necessary.

#' The following still available but they will be removed soon:

"iag_racm": "Exhaust", "Evaporative" or "Liquid".

"iag" or "iag_cb05": "Exhaust", "Evaporative" or "Liquid".

"iag_cb05v2": "Exhaust", "Evaporative" or "Liquid".

"neu_cb05": "Exhaust", "Evaporative" or "Liquid".

"petroiag_cb05": "Exhaust", "Evaporative" or "Liquid".

"petroiag_cb05v2": "Exhaust", "Evaporative" or "Liquid".

list
when TRUE returns a list with number of elements of the list as the number species of pollutants

pmpar
Numeric vector for PM speciation eg: c(e_so4i = 0.0077, e_so4j = 0.0623, e_no3i = 0.00247, e_no3j = 0.01053, e_pm25i = 0.1, e_pm25j = 0.3, e_orgi = 0.0304, e_orgj = 0.1296, e_eci = 0.056, e_ecj = 0.024, h2o = 0.277) These are default values. however, when this argument is present, new values are used.

verbose
Logical to show more information

Value

dataframe of speciation in grams or mols

Note

spec "pmiag" speciate pm2.5 into e_so4i, e_so4j, e_no3i, e_no3j, e_mp2.5i, e_mp2.5j, e_orgi, e_orgj, e_eci, e_ecj and h2o. Reference: Rafee, S.: Estudo numerico do impacto das emissoes veiculares e fixas da cidade de Manaus nas concentracoes de poluentes atmosfericos da regiao amazonica, Master thesis, Londrina: Universidade Tecnologica Federal do Parana, 2015.

specs: "neu_cb05", "pmneu" and "pmneu2" provided by Daniel Schuch, from Northeastern University

Speciation with fuels "E25", "E100" and "B5" made by Prof. Leila Martins (UTFPR), represents BRAZILIAN fuel

pmiag2 pass the mass only on j fraction

References

"bcom": Ntziachristos and Zamaras. 2016. Passneger cars, light commercial trucks, heavy-duty vehicles including buses and motor cycles. In: EEA, EMEP. EEA air pollutant emission inventory guidebook-2009. European Environment Agency, Copenhagen, 2016

"tyre", "brake" and "road": Ntziachristos and Boulter 2016. Automobile tyre and brake wear and road abrasion. In: EEA, EMEP. EEA air pollutant emission inventory guidebook-2009. European Environment Agency, Copenhagen, 2016

"iag": Ibarra-Espinosa S. Air pollution modeling in Sao Paulo using bottom-up vehicular emissions inventories. 2017. PhD thesis. Instituto de Astronomia, Geofisica e Ciencias Atmosfericas, Universidade de Sao Paulo, Sao Paulo, page 88. Speciate EPA: https://cfpub.epa.gov/speciate/. : K. Sexton, H. Westberg, "Ambient hydrocarbon and ozone measurements downwind of a large automotive painting plant" Environ. Sci. Tchnol. 14:329 (1980).P.A. Scheff, R.A. Schauer, James J., Kleeman, Mike J., Cass, Glen R., Characterization and Control of Organic Compounds Emitted from Air Pollution Sources, Final Report, Contract 93-329, prepared for California Air Resources Board Research Division, Sacramento, CA, April 1998. 2004 NPRI National Databases as of April 25, 2006, http://www.ec.gc.ca/pdb/npri/npri_dat_rep_e.cfm. Memorandum Proposed procedures for preparing composite speciation profiles using Environment Canada s National Pollutant Release Inventory (NPRI) for stationary sources, prepared by Ying Hsu and Randy Strait of E.H. Pechan Associates, Inc. for David Niemi, Marc Deslauriers, and Lisa Graham of Environment Canada, September 26, 2006.

Examples

if (FALSE) {
# Do not run
pm <- rnorm(n = 100, mean = 400, sd = 2)
(df <- speciate(pm, veh = "PC", fuel = "G", eu = "I"))
(df <- speciate(pm, spec = "brake", veh = "PC", fuel = "G", eu = "I"))
(dfa <- speciate(pm, spec = "iag", veh = "veh", fuel = "G", eu = "Exhaust"))
(dfb <- speciate(pm, spec = "iag_cb05v2", veh = "veh", fuel = "G", eu = "Exhaust"))
(dfb <- speciate(pm, spec = "neu_cb05", veh = "veh", fuel = "G", eu = "Exhaust"))
pm <- units::set_units(pm, "g/km^2/h")
(dfb <- speciate(as.data.frame(pm), spec = "pmiag", veh = "veh", fuel = "G", eu = "Exhaust"))
(dfb <- speciate(as.data.frame(pm), spec = "pmneu", veh = "veh", fuel = "G", eu = "Exhaust"))
(dfb <- speciate(as.data.frame(pm), spec = "pmneu2", veh = "veh", fuel = "G", eu = "Exhaust"))
}