DEPRECATED Construction function for class "EmissionsList"

EmissionsList returns a transformed object with class "EmissionsList".

Evaporative returns a transformed object with class "Evaporative" and units g. This class represents the daily emissions presented by Mellios G and Ntziachristos (2016) Gasoline evaporation, Tier 2. Eventually it will be incorporated the techniques of Tier 3.

This function was deprecated. If the user want some UF-based EF, run MOVES and extract EF

emis_wrf returns a dataframe with columns lat, long, id, pollutants, local time and GMT time. This dataframe has the proper format to be used with WRF assimilation system: "ASimilation System 4 WRF (AS4WRF Vera-Vala et al (2016))

The functions listed below are deprecated and will be defunct in the near future. When possible, alternative functions with similar functionality are also mentioned. Help pages for deprecated functions are available at help("-deprecated").

EmissionsList(x, ...)

Evaporative(x, ...)

ef_ive(
  description = "Auto/Sml Truck",
  fuel = "Petrol",
  weight = "Light",
  air_fuel_control = "Carburetor",
  exhaust = "None",
  evaporative = "PCV",
  mileage,
  pol,
  details = FALSE
)

emis_wrf(sdf, nr = 1, dmyhm, tz, crs = 4326, islist)

hot_soak(x, carb, p, eshotc, eswarmc, eshotfi)

running_losses(x, erhotfi, lkm, carb, p, erhotc, erwarmc)

Arguments

x

Object with class "numeric"

...

ignored

description

Character; "Auto/Sml Truck" "Truck/Bus" or"Sml Engine".

fuel

Character; "Petrol", "NG Retrofit", "Natural Gas", "Prop Retro.", "Propane", "EthOH Retrofit", "OEM Ethanol", "Diesel", "Ethanol" or "CNG/LPG".

weight

Character; "Light", "Medium", "Heavy", "Lt", "Med" or "Hvy"

air_fuel_control

Character; One of the following characters: "Carburetor", "Single-Pt FI", "Multi-Pt FI", "Carb/Mixer", "FI", "Pre-Chamber Inject.", "Direct Injection", "2-Cycle", "2-Cycle, FI", "4-Cycle, Carb", "4-Cycle, FI" "4-Cycle"

exhaust

Character: "None", "2-Way", "2-Way/EGR", "3-Way", "3-Way/EGR", "None/EGR", "LEV", "ULEV", "SULEV", "EuroI", "EuroII", "EuroIII", "EuroIV", "Hybrid", "Improved", "EGR+Improv", "Particulate", "Particulate/NOx", "EuroV", "High Tech" or "Catalyst"

evaporative

Character: "PCV", "PCV/Tank" or"None".

mileage

Numeric; mileage of vehicle by age of use km.

pol

Character; One of the following characters: "Carburetor", "Single-Pt FI", "Multi-Pt FI", "Carb/Mixer", "FI", "Pre-Chamber Inject.", "Direct Injection", "2-Cycle", "2-Cycle, FI", "4-Cycle, Carb", "4-Cycle, FI" "4-Cycle" #'

"VOC_gkm"	"CO_gkm"	"NOx_gkm"	"PM_gkm"

"Pb_gkm"	"SO2_gkm"	"NH3_gkm"	"1,3-butadiene_gkm"

"formaldehyde_gkm"	"acetaldehyde_gkm"	"benzene_gkm"	"EVAP_gkm"

"CO2_gkm"	"N20_gkm"	"CH4_gkm"	"VOC_gstart"

"CO_gstart"	"NOx_gstart"	"PM_gstart"	"Pb_gstart"

"SO2_gstart"	"NH3_gstart"	"1,3-butadiene_gstart"	"formaldehyde_gstart"

"acetaldehyde_gstart"	"benzene_gstart"	"EVAP_gstart"	"CO2_gstart"

"N20_gstart"	"CH4_gstart"

details

Logical; option to see or not more information about vehicle.

sdf

Gridded emissions, which can be a SpatialPolygonsDataFrame, or a list of SpatialPolygonsDataFrame, or a sf object of "POLYGON". The user must enter a list with 36 SpatialPolygonsDataFrame with emissions for the mechanism CBMZ.

nr

Number of repetitions of the emissions period

dmyhm

String indicating Day Month Year Hour and Minute in the format "d-m-Y H:M" e.g.: "01-05-2014 00:00" It represents the time of the first hour of emissions in Local Time

tz

Time zone as required in for function as.POSIXct

crs

Coordinate reference system, e.g: "+init=epsg:4326". Used to transform the coordinates of the output

islist

logical value to indicate if sdf is a list or not

object

Object with class "Evaporative"

Value

Objects of class "EmissionsList" and numeric elements as "units"

Objects of class "Evaporative" or "units"

An emission factor by annual mileage.

Details

ef_ive returns the base emission factors from the the IVE model. This function depend on vectorized mileage, which means your can enter with the mileage by age of use and the name of the pollutant.

Evaporative

For Evaporative, use emis_evap.

For running_losses, use emis_evap.

For running_losses, use emis_evap.

References

Nicole Davis, James Lents, Mauricio Osses, Nick Nikkila, Matthew Barth. 2005. Development and Application of an International Vehicle Emissions Model. Transportation Research Board, 81st Annual Meeting, January 2005, Washington, D.C.

Mellios G and Ntziachristos 2016. Gasoline evaporation. In: EEA, EMEP. EEA air pollutant emission inventory guidebook-2009. European Environment Agency, Copenhagen, 2009

Mellios G and Ntziachristos 2016. Gasoline evaporation. In: EEA, EMEP. EEA air pollutant emission inventory guidebook-2009. European Environment Agency, Copenhagen, 2009

See also

Examples

if (FALSE) {
# do not run
# DEPRECATED
}
if (FALSE) {
# do not run
# DEPRECATED
}
if (FALSE) {
# Do not run
# Passenger Cars PC
data(fkm)
# cumulative mileage from 1 to 50 years of use, 40:50
mil <- cumsum(fkm$KM_PC_E25(1:50))
ef_ive("Truck/Bus", mileage = mil, pol = "CO_gkm")
ef_ive(mileage = mil, pol = "CO_gkm", details = TRUE)
}
if (FALSE) {
# do not run
# Use library(eixport)
# eixport::to_as4wrf()
}
if (FALSE) {
# Do not run
}
if (FALSE) {
# Do not run
}