DEPRECATED Construction function for class "EmissionsList"

EmissionsList returns a tranformed object with class "EmissionsList".

Evaporative returns a tranformed object with class "Evaporative" and units g. This class represents the daily emissions presented by Mellios G and Ntziachristos (2016) Gasoline evaporation, Tier 2. Eventually it will be incorporated the techniques of Tier 3.

emis_wrf returns a dataframes with columns lat, long, id, pollutants, local time and GMT time. This dataframe has the proper format to be used with WRF assimilation system: "ASimilation System 4 WRF (AS4WRF Vera-Vala et al (2016))

The functions listed below are deprecated and will be defunct in the near future. When possible, alternative functions with similar functionality are also mentioned. Help pages for deprecated functions are available at help("-deprecated").

EmissionsList(x, ...)

Evaporative(x, ...)

emis_wrf(sdf, nr = 1, dmyhm, tz, crs = 4326, islist)

hot_soak(x, carb, p, eshotc, eswarmc, eshotfi)

running_losses(x, erhotfi, lkm, carb, p, erhotc, erwarmc)

Arguments

x
Object with class "numeric"

...
ignored

sdf
Gridded emissions, which can be a SpatialPolygonsDataFrame, or a list of SpatialPolygonsDataFrame, or a sf object of "POLYGON". The user must enter a list with 36 SpatialPolygonsDataFrame with emissions for the mechanism CBMZ.

nr
Number of repetitions of the emissions period

dmyhm
String indicating Day Month Year Hour and Minute in the format "d-m-Y H:M" e.g.: "01-05-2014 00:00" It represents the time of the first hour of emissions in Local Time

tz
Time zone as required in for function as.POSIXct

crs
Coordinate reference system, e.g: "+init=epsg:4326". Used to transform the coordinates of the output

islist
logical value to indicate if sdf is a list or not

object
Object with class "Evaporative"

Value

Objects of class "EmissionsList" and numeric elements as "units"

Objects of class "Evaporative" or "units"

Evaporative

For Evaporative, use emis_evap.

For running_losses, use emis_evap.

For running_losses, use emis_evap.

References

Mellios G and Ntziachristos 2016. Gasoline evaporation. In: EEA, EMEP. EEA air pollutant emission inventory guidebook-2009. European Environment Agency, Copenhagen, 2009

Mellios G and Ntziachristos 2016. Gasoline evaporation. In: EEA, EMEP. EEA air pollutant emission inventory guidebook-2009. European Environment Agency, Copenhagen, 2009

See also

Examples

if (FALSE) {
# do not run
# DEPRECATED
}
if (FALSE) {
# do not run
# DEPRECATED
}
if (FALSE) {
# do not run
# Use library(eixport)
# eixport::to_as4wrf()
}
if (FALSE) {
# Do not run
}
if (FALSE) {
# Do not run
}